ready for test

+++ controllers/controller-expense.php
@@ -1,10 +1,8 @@
 <?php 
 session_start();
-
 include_once "../models/Expense.php";
 include_once "../models/Worker.php";
-
-
+// Change the selected link in the aside menu
 $page_name = [];
 $page_name["expense"] = "selected-aside";
 $connected = false;
@@ -18,10 +16,12 @@ if(!isset($_SESSION["id"])){
 
 if($_SERVER["REQUEST_METHOD"] == "POST")
 {
-
     if(isset($_POST["delete"]))
     {
-        Expense::deleteExpense($_POST["delete"]) ;
+        if(Expense::verifyWorkerIDforExpense($_POST["delete"],$_SESSION["id"]))
+        {
+        Expense::deleteExpense($_POST["delete"]) ;   
+        }
         header("Location: controller-expense.php");
     }
 }
